opt: some cleanup in CTE code
This commit cleans up the optbuilder code related to CTEs and improves
its documentation.

Release note: None
