+++ pkg/sql/opt/optbuilder/with.go
@@ -11,15 +11,70 @@
 package optbuilder
 
 import (
+	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/props"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/props/physical"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sqltelemetry"
 	"github.com/cockroachdb/errors"
 )
 
+// pushWithFrame pushes a new frame onto the CTE stack.
+//
+// The CTE stack is used to "hoist" CTEs to an appropriate boundary before
+// constructing the With operators. For run-of-the-mill queries, there are no
+// boundaries and all CTEs are hoisted at the top level. However, certain
+// statements (e.g. EXPLAIN, CREATE VIEW, CREATE TABLE AS) must introduce a
+// boundary.
+func (b *Builder) pushWithFrame() {
+	b.cteStack = append(b.cteStack, []cteSource{})
+}
+
+// addCTEToWithFrame adds a CTE to the current frame.
+func (b *Builder) addCTEToWithFrame(cte cteSource) {
+	level := len(b.cteStack) - 1
+	b.cteStack[level] = append(b.cteStack[level], cte)
+}
+
+// popWithFrame pops a frame from the CTE stack and wraps the given scope's
+// expression with With operators for the expressions in this frame.
+func (b *Builder) popWithFrame(s *scope) {
+	level := len(b.cteStack) - 1
+	ctes := b.cteStack[level]
+	b.cteStack = b.cteStack[:level]
+
+	s.expr = b.buildWiths(s.expr, ctes)
+}
+
+// buildWiths adds With expressions on top of an expression.
+func (b *Builder) buildWiths(expr memo.RelExpr, ctes []cteSource) memo.RelExpr {
+	if len(ctes) == 0 {
+		return expr
+	}
+
+	// Since later CTEs can refer to earlier ones, we want to add these in
+	// reverse order.
+	for i := len(ctes) - 1; i >= 0; i-- {
+		expr = b.factory.ConstructWith(
+			ctes[i].expr,
+			expr,
+			&memo.WithPrivate{
+				ID:           ctes[i].id,
+				Name:         string(ctes[i].name.Alias),
+				Mtr:          ctes[i].mtr,
+				OriginalExpr: ctes[i].originalExpr,
+			},
+		)
+	}
+	return expr
+}
+
+// processWiths is used when building a statement that has a WITH clause. It
+// builds any CTEs defined by the WITH clause and calls the given function to
+// build the statement itself.
 func (b *Builder) processWiths(
 	with *tree.With, inScope *scope, buildStmt func(inScope *scope) *scope,
 ) *scope {
@@ -31,6 +86,75 @@ func (b *Builder) processWiths(
 	return outScope
 }
 
+// buildCTEs constructs expressions for the CTEs defined by a WITH clause and
+// adds them to the CTE stack.
+func (b *Builder) buildCTEs(with *tree.With, inScope *scope) (outScope *scope) {
+	if with == nil {
+		return inScope
+	}
+
+	outScope = inScope.push()
+	addedCTEs := make([]cteSource, len(with.CTEList))
+	hasRecursive := false
+
+	// Make a fake subquery to ensure that no CTEs are correlated.
+	// TODO(justin): relax this restriction.
+	outer := b.subquery
+	defer func() { b.subquery = outer }()
+	b.subquery = &subquery{}
+
+	outScope.ctes = make(map[string]*cteSource)
+	for i, cte := range with.CTEList {
+		hasRecursive = hasRecursive || with.Recursive
+		cteExpr, cteCols := b.buildCTE(cte, outScope, with.Recursive)
+
+		// TODO(justin): lift this restriction when possible. WITH should be hoistable.
+		if b.subquery != nil && !b.subquery.outerCols.Empty() {
+			panic(pgerror.Newf(pgcode.FeatureNotSupported, "CTEs may not be correlated"))
+		}
+
+		aliasStr := cte.Name.Alias.String()
+		if _, ok := outScope.ctes[aliasStr]; ok {
+			panic(pgerror.Newf(
+				pgcode.DuplicateAlias, "WITH query name %s specified more than once", aliasStr,
+			))
+		}
+
+		id := b.factory.Memo().NextWithID()
+		b.factory.Metadata().AddWithBinding(id, cteExpr)
+
+		addedCTEs[i] = cteSource{
+			name:         cte.Name,
+			cols:         cteCols,
+			originalExpr: cte.Stmt,
+			expr:         cteExpr,
+			id:           id,
+			mtr:          cte.Mtr,
+		}
+		cte := &addedCTEs[i]
+		outScope.ctes[cte.name.Alias.String()] = cte
+		b.addCTEToWithFrame(*cte)
+
+		if cteExpr.Relational().CanMutate && !inScope.atRoot {
+			panic(
+				pgerror.Newf(
+					pgcode.FeatureNotSupported,
+					"WITH clause containing a data-modifying statement must be at the top level",
+				),
+			)
+		}
+
+	}
+
+	telemetry.Inc(sqltelemetry.CteUseCounter)
+	if hasRecursive {
+		telemetry.Inc(sqltelemetry.RecursiveCteUseCounter)
+	}
+
+	return outScope
+}
+
+// buildCTE constructs an expression for a CTE.
 func (b *Builder) buildCTE(
 	cte *tree.CTE, inScope *scope, isRecursive bool,
 ) (memo.RelExpr, physical.Presentation) {
